scripts/webify/pyproject.toml
[tool.poetry]
name = "webify"
version = "0.1.0"
description = ""
authors = ["Faris Chebib <faris@theluckybead.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.5.0"
tqdm = "^4.65.0"
moto = "^4.1.12"
aioboto3 = "^11.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
